better board debug with colored output

+++ src/game/constants.rs
@@ -1,4 +1,4 @@
-pub const ROW_SIZE: i8 = 3;
+pub const ROW_SIZE: usize = 3;
 pub const BOARD_SIZE: usize = 9;
 
 pub const CARD_SIDES: usize = 4;
